reworking projects page
